requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
ament-clang-format==0.12.7
ament-clang-tidy==0.12.7
ament-cmake-google-benchmark==1.3.5
ament-cmake-test==1.3.5
ament-copyright==0.12.7
ament-cppcheck==0.12.7
ament-cpplint==0.12.7
ament-flake8==0.12.7
ament-index-python==1.4.0
ament-lint==0.12.7
ament-lint-cmake==0.12.7
ament-mypy==0.12.7
ament-package==0.14.0
ament-pclint==0.12.7
ament-pep257==0.12.7
ament-pycodestyle==0.12.7
ament-pyflakes==0.12.7
ament-uncrustify==0.12.7
ament-xmllint==0.12.7
blinker==1.6.2
builtin-interfaces==1.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
composition-interfaces==1.2.1
demo-nodes-py==0.20.3
diagnostic-msgs==4.2.3
diffusers==0.21.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-guard-conditions==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
examples-rclpy-pointcloud-publisher==0.15.1
examples-tf2-py==0.25.3
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Executor==1.0.0
Flask-JWT-Extended==4.5.2
fsspec==2023.9.0
ftfy==6.1.1
geometry-msgs==4.2.3
huggingface-hub==0.17.1
idna==3.4
importlib-metadata==6.8.0
interactive-markers==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
laser-geometry==2.4.0
launch==1.0.4
launch-pytest==1.0.4
launch-ros==0.19.5
launch-testing==1.0.4
launch-testing-examples==0.15.1
launch-testing-ros==0.19.5
launch-xml==1.0.4
launch-yaml==1.0.4
lifecycle-msgs==1.2.1
lifecycle-py==0.20.3
lit==16.0.6
logging-demo==0.20.3
map-msgs==2.1.0
MarkupSafe==2.1.3
message-filters==4.3.3
mpmath==1.3.0
nav-msgs==4.2.3
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open-clip-torch==2.20.0
opencv-python==4.8.0.76
osrf-pycommon==2.1.3
packaging==23.1
pendulum-msgs==0.20.3
Pillow==10.0.0
protobuf==3.20.3
PyJWT==2.8.0
python-dotenv==1.0.0
python-qt-binding==1.1.1
PyYAML==6.0.1
qt-dotgraph==2.2.2
qt-gui==2.2.2
qt-gui-cpp==2.2.2
qt-gui-py-common==2.2.2
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.9
rcutils==5.1.3
regex==2023.8.8
requests==2.31.0
resource-retriever==3.1.1
rmw-dds-common==1.6.0
ros2action==0.18.7
ros2bag==0.15.7
ros2cli==0.18.7
ros2cli-test-interfaces==0.18.7
ros2component==0.18.7
ros2doctor==0.18.7
ros2interface==0.18.7
ros2launch==0.19.5
ros2lifecycle==0.18.7
ros2multicast==0.18.7
ros2node==0.18.7
ros2param==0.18.7
ros2pkg==0.18.7
ros2run==0.18.7
ros2service==0.18.7
ros2test==0.4.0
ros2topic==0.18.7
ros2trace==4.1.1
rosbag2-interfaces==0.15.7
rosbag2-py==0.15.7
rosbag2-storage-mcap-testdata==0.15.7
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-dds-idl==0.8.1
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.1
rosidl-typesupport-fastrtps-cpp==2.2.1
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt==1.1.5
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.2
rqt-graph==1.3.0
rqt-gui==1.1.5
rqt-gui-py==1.1.5
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.5
rqt-py-console==1.0.2
rqt-reconfigure==1.1.1
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
safetensors==0.3.3
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
sentencepiece==0.1.99
shape-msgs==4.2.3
sros2==0.10.4
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.12
test-launch-ros==0.19.5
test-msgs==1.2.1
test-tracetools-launch==4.1.1
tf2-geometry-msgs==0.25.3
tf2-kdl==0.25.3
tf2-msgs==0.25.3
tf2-py==0.25.3
tf2-ros-py==0.25.3
tf2-tools==0.25.3
timm==0.9.7
tokenizers==0.13.3
topic-monitor==0.20.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
tracetools-launch==4.1.1
tracetools-read==4.1.1
tracetools-test==4.1.1
tracetools-trace==4.1.1
trajectory-msgs==4.2.3
transformers==4.33.1
triton==2.0.0
turtlesim==1.4.2
typing_extensions==4.7.1
unique-identifier-msgs==2.2.1
urllib3==2.0.4
visualization-msgs==4.2.3
wcwidth==0.2.6
Werkzeug==2.3.7
zipp==3.16.2
